third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/script-tag-with-source-data-url-expected.txt

CONSOLE ERROR: line 5: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Cscript%20src=%22data:,alert(1)%22' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
